LoL channels creator added

+++ src/main/java/events/LoLEvent.java
@@ -20,13 +20,15 @@ public class LoLEvent extends ListenerAdapter {
     {
         Guild guild = event.getMember().getGuild();
         String name;
-        if (event.getChannelJoined().getId().compareTo("797300475237498890")==0) {
+        if (event.getChannelJoined().getId().equals("797300475237498890")) {
             boolean found=true;
             int i;
             for(i=1;found;i++)
             {
+                System.out.println(i);
                 found=ChannelNames.lol.contains(i-1);
             }
+            System.out.println(i);
             ChannelNames.lol.add(i-1);
             name="「\uD83D\uDD08」 LoL #"+ (i-1);
                 ChannelAction<VoiceChannel> ca = guild.createVoiceChannel(name);
@@ -42,13 +44,15 @@ public class LoLEvent extends ListenerAdapter {
     public void onGuildVoiceMove(GuildVoiceMoveEvent event) {
         Guild guild = event.getMember().getGuild();
         String name;
-        if (event.getChannelJoined().getId().compareTo("797300475237498890")==0) {
+        if (event.getChannelJoined().getId().equals("797300475237498890")) {
             boolean found=true;
             int i;
             for(i=1;found;i++)
             {
+                System.out.println(i);
                 found=ChannelNames.lol.contains(i-1);
             }
+            System.out.println(i);
             ChannelNames.lol.add(i-1);
             name="「\uD83D\uDD08」 LoL #"+ (i-1);
                 ChannelAction<VoiceChannel> ca = guild.createVoiceChannel(name);
@@ -59,22 +63,22 @@ public class LoLEvent extends ListenerAdapter {
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
-        }else if(Objects.requireNonNull(event.getChannelLeft().getParent()).getId().equals("797297810139250728"))
+        }else if(Objects.requireNonNull(event.getChannelLeft().getParent()).getId().equals("797297810139250728")&&!event.getChannelLeft().getId().equals("797300475237498890"))
         {
             if(event.getChannelLeft().getMembers().isEmpty())
             {
-                ChannelNames.lol.remove(Integer.parseInt(String.valueOf(event.getChannelLeft().getName().charAt(event.getChannelLeft().getName().length()-1))));
+                ChannelNames.lol.remove(Integer.parseInt(event.getChannelLeft().getName().substring(event.getChannelLeft().getName().length()-1)));
                 event.getChannelLeft().delete().queue();
             }
         }
     }
     public void onGuildVoiceLeave(GuildVoiceLeaveEvent event)
     {
-        if(Objects.requireNonNull(event.getChannelLeft().getParent()).getId().equals("797297810139250728"))
+        if(Objects.requireNonNull(event.getChannelLeft().getParent()).getId().equals("797297810139250728")&&!event.getChannelLeft().getId().equals("797300475237498890"))
         {
             if(event.getChannelLeft().getMembers().isEmpty())
             {
-                ChannelNames.lol.remove(Integer.parseInt(String.valueOf(event.getChannelLeft().getName().charAt(event.getChannelLeft().getName().length()-1))));
+                ChannelNames.lol.remove(Integer.parseInt(event.getChannelLeft().getName().substring(event.getChannelLeft().getName().length()-1)));
                 event.getChannelLeft().delete().queue();
             }
         }
